requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
frozenlist==1.6.0
idna==3.10
iso8601==2.1.0
multidict==6.4.3
propcache==0.3.1
python-dotenv==1.1.0
twitchio==2.10.0
typing_extensions==4.13.2
yarl==1.20.0
